+++ src/components/Button/Button.tsx
@@ -1,22 +1,6 @@
-import { Text, TouchableOpacity, TouchableOpacityProps } from 'react-native';
-import { styles } from './style';
+import { StyleSheet, Text, TouchableOpacity, TouchableOpacityProps } from 'react-native';
 import { theme } from '~/lib/theme';
-
-type ButtonProps = {
-  onPress?: TouchableOpacityProps['onPress'];
-  textColor?: string;
-  title?: string;
-} & TouchableOpacityProps;
-
-type FilledButtonProps = {
-  backgroundColor?: string;
-} & ButtonProps
-
-type OutlineButtonProps = {
-  outlineColor?: string;
-} & ButtonProps & FilledButtonProps;
-
-type TransparentButtonProps = {} & ButtonProps
+import { FilledButtonProps, OutlineButtonProps, TransparentButtonProps } from './button-props';
 
 class Button {
   Filled({ backgroundColor = theme.colors.primary, textColor = theme.colors.white, title, onPress }: FilledButtonProps) {
@@ -44,4 +28,24 @@ class Button {
   }
 }
 
+const styles = StyleSheet.create({
+  button: {
+    alignItems: 'center',
+    backgroundColor: 'transparent',
+    borderColor: 'transparent',
+    borderRadius: 6,
+    borderStyle: 'solid',
+    borderWidth: 1,
+    flexDirection: 'row',
+    justifyContent: 'center',
+    padding: 16,
+  },
+  buttonText: {
+    fontSize: theme.fonts.sizes.base,
+    fontWeight: '600',
+    textAlign: 'center',
+  },
+});
+
 export default new Button;
+
