--- src/screens/Access/styles.ts
@@ -1,11 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        gap: 25,
-        padding: 25,
-    },
-});
-
-export default styles;
