--- src/screens/signup/styles.ts
@@ -1,9 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        padding: 25,
-    }
-})
